Add adapter and databases

+++ app/src/main/java/com/example/databasetesting/ErrorMessage.kt
@@ -2,11 +2,12 @@ package com.example.databasetesting
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import org.w3c.dom.Text
 
 @Entity(tableName = "errorMessageTable")
 data class ErrorMessage(
     @PrimaryKey
-    val id: Int,
+    val id: Long,
     val errorTitle: String,
     val errorDescription: String
 )
